requirements.txt
Pillow==6.2.0
python-telegram-bot==11.1.0
pyfiglet==0.8.post1